requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
black==22.12.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
distlib==0.3.6
exceptiongroup==1.1.0
filelock==3.9.0
frozenlist==1.3.3
identify==2.5.18
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
mypy==0.991
mypy-extensions==0.4.3
nodeenv==1.7.0
packaging==23.0
pathspec==0.10.3
Pillow==9.4.0
platformdirs==2.6.2
pluggy==1.0.0
pre-commit==3.0.4
pytest==7.2.1
pytest-mock==3.10.0
PyYAML==6.0
requests==2.28.2
ruff==0.0.246
tomli==2.0.1
types-requests==2.28.11.8
types-urllib3==1.26.25.4
typing_extensions==4.4.0
urllib3==1.26.14
virtualenv==20.19.0
yarl==1.8.2
